src/main/java/com/cib/back/api/controllers/MessageController.java
package com.cib.back.api.controllers;

import com.cib.back.api.models.Response;
import com.cib.back.api.services.MessageService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("api/messages/")
public class MessageController {
    private final MessageService messageService;

    @Autowired
    public MessageController(MessageService messageService) {
        this.messageService = messageService;
    }

    @GetMapping()
    public ResponseEntity<Response> getAllMessage() {
        Response response = Response.from(messageService.getAllMessage());

        return ResponseEntity.ok(response);
    }
}
